Remove socket timeouts and add operation type to progress broadcasts
This commit removes connection and server accept timeouts in `FileTransferService` to allow for unlimited time when transferring large files. It also adds a new extra, `EXTRA_OPERATION_TYPE`, to progress broadcasts to indicate the current phase of the file transfer (e.g., "encrypting", "transferring", "decrypting").

The `WiFiTransferHelper` was also updated to remove the `fileServerSocket` as it's no longer managed by this class.